requirements.txt
requests
pandas
python-slugify